Generic Jetbrains product Debian package generator.
Reworked the Intellij IDEA repository to be a generic Jetbrains
product package generator installer.

Jetbrains has aligned all their products around the same EAP and
installation structure so it is fairly easy to create a single
generator for all of them. Some product specific hacks are
implemented.

Currently Clion and IDEA are supported but more products are (fairly)
easy to add now.

This is a fork of the public repository for IDEA Intellij [1] and
merge of a few local forks that I have maintained to build other
products.

[1]: https://github.com/trygvis/intellij-idea-dpkg